% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{find_na_rows}
\alias{find_na_rows}
\title{Find all rows with an NA in any column}
\usage{
find_na_rows(.data)
}
\arguments{
\item{.data}{A data frame}
}
\value{
A subset of the data that contains all rows with an NA in any column.
}
\description{
Find all rows with an NA in any column
}
\examples{
na_car <- mtcars
na_car[which(na_car[,"disp"] < 160), "disp"] <- NA
na_car[which(na_car[, "hp"] == 110), "hp"] <- NA

find_na_rows(.data = na_car)

}
